# ENH: speed up for differential_evolution updating='deferred'