--- a/scipy/optimize/_differentialevolution.py
+++ b/scipy/optimize/_differentialevolution.py
@@ -7,7 +7,8 @@ import warnings
 import numpy as np
 from scipy.optimize import OptimizeResult, minimize
 from scipy.optimize._optimize import _status_message, _wrap_callback
-from scipy._lib._util import check_random_state, MapWrapper, _FunctionWrapper
+from scipy._lib._util import (check_random_state, MapWrapper, _FunctionWrapper,
+                              rng_integers)
 
 from scipy.optimize._constraints import (Bounds, new_bounds_to_old,
                                          NonlinearConstraint, LinearConstraint)
@@ -1567,7 +1568,7 @@ class DifferentialEvolutionSolver:
                 # number of candidates that were mutated.
                 # For `immediate` S = 1. For `deferred` this will
                 # probably be the entire population, S = num_population_members
-                parameters = self._scale_parameters(trial)[0]
+                parameters = self._scale_parameters(trial)
 
                 # determine the energy of the objective function
                 if self._wrapped_constraints:
@@ -1610,7 +1611,9 @@ class DifferentialEvolutionSolver:
 
             # 'deferred' approach, vectorised form.
             # create trial solutions
-            trial_pop = self._mutate(np.arange(self.num_population_members))
+            trial_pop = self._mutate_many(
+                np.arange(self.num_population_members)
+            )
 
             # enforce bounds
             self._ensure_constraint(trial_pop)
@@ -1668,39 +1671,45 @@ class DifferentialEvolutionSolver:
         if oob := np.count_nonzero(mask):
             trial[mask] = self.random_number_generator.uniform(size=oob)
 
-    def _mutate(self, candidate):
-        """Create a trial vector based on a mutation strategy."""
+    def _mutate_custom(self, candidate):
         rng = self.random_number_generator
-
-        candidate = np.atleast_1d(candidate)
-        S = len(candidate)
-
-        if callable(self.strategy):
-            _population = self._scale_parameters(self.population)
+        msg = (
+            "strategy must have signature"
+            " f(candidate: int, population: np.ndarray, rng=None) returning an"
+            " array of shape (N,)"
+        )
+        _population = self._scale_parameters(self.population)
+        if not len(np.shape(candidate)):
+            # single entry in population
+            trial = self.strategy(candidate, _population, rng=rng)
+            if trial.shape != (self.parameter_count,):
+                raise RuntimeError(msg)
+        else:
             trial = np.array(
                 [self.strategy(c, _population, rng=rng) for c in candidate],
                 dtype=float
             )
             if trial.shape != (S, self.parameter_count):
-                raise RuntimeError(
-                    "strategy must have signature"
-                    " f(candidate: int, population: np.ndarray, rng=None)"
-                    " returning an array of shape (N,)"
-                )
-            return self._unscale_parameters(trial)
+                raise RuntimeError(msg)
+        return self._unscale_parameters(trial)
 
-        # trial should be at least 2-D, (S, N), because len(candidate) >= 1
-        trial = np.copy(self.population[candidate])
-        fill_point = rng.choice(self.parameter_count, size=S)
+    def _mutate_many(self, candidates):
+        """Create trial vectors based on a mutation strategy."""
+        rng = self.random_number_generator
 
-        # (S, 5)
-        samples = np.array([self._select_samples(c, 5) for c in candidate])
+        S = len(candidates)
+        if callable(self.strategy):
+            return self._mutate_custom(candidates)
+
+        trial = np.copy(self.population[candidates])
+        samples = np.array([self._select_samples(c, 5) for c in candidates])
 
         if self.strategy in ['currenttobest1exp', 'currenttobest1bin']:
-            bprime = self.mutation_func(candidate, samples)
+            bprime = self.mutation_func(candidates, samples)
         else:
             bprime = self.mutation_func(samples)
 
+        fill_point = rng_integers(rng, self.parameter_count, size=S)
         crossovers = rng.uniform(size=(S, self.parameter_count))
         crossovers = crossovers < self.cross_over_probability
         if self.strategy in self._binomial:
@@ -1708,13 +1717,13 @@ class DifferentialEvolutionSolver:
             # If you fill in modulo with a loop you have to set the last one to
             # true. If you don't use a loop then you can have any random entry
             # be True.
-            idx = np.arange(S)
-            crossovers[idx, fill_point[idx]] = True
+            i = np.arange(S)
+            crossovers[i, fill_point[i]] = True
             trial = np.where(crossovers, bprime, trial)
             return trial
 
         elif self.strategy in self._exponential:
-            crossovers[:, 0] = True
+            crossovers[..., 0] = True
             for j in range(S):
                 i = 0
                 init_fill = fill_point[j]
@@ -1725,22 +1734,62 @@ class DifferentialEvolutionSolver:
 
             return trial
 
+    def _mutate(self, candidate):
+        """Create a trial vector based on a mutation strategy."""
+        rng = self.random_number_generator
+
+        if callable(self.strategy):
+            return self._mutate_custom(candidate)
+
+        fill_point = rng_integers(rng, self.parameter_count)
+        samples = self._select_samples(candidate, 5)
+
+        trial = np.copy(self.population[candidate])
+
+        if self.strategy in ['currenttobest1exp', 'currenttobest1bin']:
+            bprime = self.mutation_func(candidate, samples)
+        else:
+            bprime = self.mutation_func(samples)
+
+        crossovers = rng.uniform(size=self.parameter_count)
+        crossovers = crossovers < self.cross_over_probability
+        if self.strategy in self._binomial:
+            # the last one is always from the bprime vector for binomial
+            # If you fill in modulo with a loop you have to set the last one to
+            # true. If you don't use a loop then you can have any random entry
+            # be True.
+            crossovers[fill_point] = True
+            trial = np.where(crossovers, bprime, trial)
+            return trial
+
+        elif self.strategy in self._exponential:
+            i = 0
+            crossovers[0] = True
+            while i < self.parameter_count and crossovers[i]:
+                trial[fill_point] = bprime[fill_point]
+                fill_point = (fill_point + 1) % self.parameter_count
+                i += 1
+
+            return trial
+
     def _best1(self, samples):
         """best1bin, best1exp"""
         # samples.shape == (S, 5)
-        r0, r1 = samples.T[:2]
+        # or
+        # samples.shape(5,)
+        r0, r1 = samples[..., :2].T
         return (self.population[0] + self.scale *
                 (self.population[r0] - self.population[r1]))
 
     def _rand1(self, samples):
         """rand1bin, rand1exp"""
-        r0, r1, r2 = samples.T[:3]
+        r0, r1, r2 = samples[..., :3].T
         return (self.population[r0] + self.scale *
                 (self.population[r1] - self.population[r2]))
 
     def _randtobest1(self, samples):
         """randtobest1bin, randtobest1exp"""
-        r0, r1, r2 = samples.T[:3]
+        r0, r1, r2 = samples[..., :3].T
         bprime = np.copy(self.population[r0])
         bprime += self.scale * (self.population[0] - bprime)
         bprime += self.scale * (self.population[r1] -
@@ -1749,7 +1798,7 @@ class DifferentialEvolutionSolver:
 
     def _currenttobest1(self, candidate, samples):
         """currenttobest1bin, currenttobest1exp"""
-        r0, r1 = samples.T[:2]
+        r0, r1 = samples[..., :2].T
         bprime = (self.population[candidate] + self.scale *
                   (self.population[0] - self.population[candidate] +
                    self.population[r0] - self.population[r1]))
@@ -1757,7 +1806,7 @@ class DifferentialEvolutionSolver:
 
     def _best2(self, samples):
         """best2bin, best2exp"""
-        r0, r1, r2, r3 = samples.T[:4]
+        r0, r1, r2, r3 = samples[..., :4].T
         bprime = (self.population[0] + self.scale *
                   (self.population[r0] + self.population[r1] -
                    self.population[r2] - self.population[r3]))
@@ -1766,7 +1815,7 @@ class DifferentialEvolutionSolver:
 
     def _rand2(self, samples):
         """rand2bin, rand2exp"""
-        r0, r1, r2, r3, r4 = samples.T[:5]
+        r0, r1, r2, r3, r4 = samples[..., :5].T
         bprime = (self.population[r0] + self.scale *
                   (self.population[r1] + self.population[r2] -
                    self.population[r3] - self.population[r4]))