This PR reduces the overhead of the differential evolution solver by vectorising the `optimize._differential_evolution.DifferentialEvolutionSolver._mutate` method. This will only have an effect when `updating='deferred'` because the entire population is mutated at once. When `_mutate='immediate` mutation still occurs for each candidate individually. For the latter case solver speed is no faster.

Note: that bitwise equivalence of solver pathway is not maintained before/after this PR. This is because the random number streams are consumed in different orders. 
i.e. setting a seed still results in a reproducible minimisation. However, using the same seed before/after this PR will not give a bitwise equivalent solution. I believe that's an acceptable tradeoff for a faster solve.

For this example:
```python
from scipy.optimize import differential_evolution
import numpy as np

N = 10
bounds=[(-10, 10)]*N

%timeit differential_evolution(rosen, bounds, updating='deferred', vectorized=True)
```

Before PR:

`2.1 s ± 24.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

After PR:

`623 ms ± 3.82 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

@mdhaber this should improve performance in the areas of stats where DE is used.