@mdhaber 2024-05-22
Thanks. I'd be willing to compare before and after performance (and with `'immediate'`/`'deferred'`) on the fit tests. Is it easy to run the benchmarks before/after?