@andyfaff 2024-05-22
> on the fit tests. Is it easy to run the benchmarks before/after?

Which benchmarks? If you're talking about the global benchmark suite, then that uses the default kwds, which won't show any difference. To see the time improvements offered by this PR one has to use `updating='deferred'`, not the default `updating='immediate'`.