@andyfaff 2024-05-23
I ran the following against this PR and against main to benchmark (couldn't figure out how to run benchmark suite easily for only a single solver):

```python
import benchmarks.go_benchmark_functions as gbf
import inspect
import time
from scipy.optimize import differential_evolution
import tqdm
import pickle

funcs = dict([
    item for item in inspect.getmembers(gbf, inspect.isclass)
    if (issubclass(item[1], gbf.Benchmark) and
        item[0] not in ('Benchmark') and
        not item[0].startswith('Problem'))
])

NITER = 10
results = {}
for fstr, func in tqdm.tqdm(funcs.items(), position=0, total=len(funcs), leave=True):
    klass = func()
    start = time.time()
    for i in range(NITER):
        differential_evolution(klass.fun, klass.bounds, updating='deferred')    
    stop = time.time()
    results[fstr] = (stop - start)/NITER
```

The results:
![time](https://github.com/scipy/scipy/assets/702934/4daa5e11-b9e6-4822-864b-27cd45e34202)

So if one is using deferred the minimisation is twice as fast. I didn't check success because that proportion shouldn't change