@dschmitz89 2024-05-23
I cannot comment on the internals of the solver but as this PR is a pure refactoring, I would trust Andrew's judgement here. Feel free to self merge.