@andyfaff 2024-05-24
Moved to draft, investigating immediate vs immediate performance indicated a slow down.