@andyfaff 2024-05-27
I split the code out into two different code paths. One that deals with mutating one candidate and one that deals with mutating the entire population. I couldn't see an easy and performant way of having them in one method. All tests pass, and I'm happy with the changes.